% by okada_n 120328
% charset=utf-8
\encoding{utf-8}
\name{plot.barplot3d}
\alias{plot.barplot3d2}
\alias{plot.barplot3d}
\title{
[Plot] Barplot 3D
}
\description{
http://addictedtor.free.fr/graphiques/graphcode.php?graph=161
}
\usage{
plot.barplot3d2 (mat, row, size = 3, space = 0.5, col.lab = NULL, row.lab = NULL, z.lab = NULL, col = "aaaaaa", grid = "white", \dots)
plot.barplot3d (heights, rows, transp = "f0", theta = 25, phi = 25, bar.size = 3, bar.space = 0.5, col.lab = NULL, row.lab = NULL, z.lab = NULL, col.bar = c("44ff58", "5844ff", "ff5844"), grid = "white", \dots, ver = 1)
}
